[WIP] Add basic support for displaying emotes
Not really upstreamable as it is; we'd need the username to be in-line
with the message if we are to follow the way element shows the messages.

Sending emotes requires more changes. Namely better support for
commands. Namely that they're not tied to the completion framework for
activation, but hooked on message being sent and starting with a `/`.