Add support for displaying redacted messages
This re-orders `genMessage` a bit to provide the basic necessary
building blocks for a redacted message (namely, header and avatar).

In turn, this uses a slightly translucent style to better represent a
redacted message.

There is a minor limitation in the current design. Assuming the
following sequence of events.

```
|¯¯| Bob
|__| Hello there!

|¯¯| Alice
|__| *Message deleted*
     Hi everyone
     Bla bla bla

|¯¯| Bob
|__| Oh hey, bla bla
```

The header and avatar for Alice will be shown with transparency since
they are attached to the redacted message.

A quick workaround would be to apply the transparency only to the
redacted *contents*. This is likely the desired solution.

A less quick solution would be to (somehow) check if there are follow-up
events and only style the header/avatar if they are all redacted events.

A less good solution would be attach a new header/avatar after a message
redaction. This is less good as it will use more vertical real-estate.