Suggest all commands when using `/`
This implements a suggestion, that I am in agreement with. Knowing which
commands exists when first using Cinny is cumbersome, as completion
requires at least one letter to be used. I had to poke at arbitrary
letters to see if I was missing any commands at first.

By always suggesting all commands initially it removes the awkwardness
of discovery. At the cost of losing the activation message.

Fixes #157.